Parameter for the effective minimum distance between embedded points

Used in embed::step_umap().

min_dist(range = c(-4, 0), trans = log10_trans())

Arguments

range
A two-element vector holding the defaults for the smallest and largest possible values, respectively.

trans
A trans object from the scales package, such as scales::log10_trans() or scales::reciprocal_trans(). If not provided, the default is used which matches the units used in range. If no transformation, NULL.

Examples

min_dist()

#> Min Distance between Points  (quantitative)
#> Transformer:  log-10 
#> Range (transformed scale): [-4, 0]