Parameter objects related to miscellaneous models.

These are objects that can be used for modeling, especially in conjunction with the parsnip package.

weight_func

surv_dist

Laplace

neighbors

Format

An object of class qual_param (inherits from param) of length 4.

Value

Each object is generated by either new_quant_param or new_qual_param.

Details

These objects are pre-made parameter sets that are useful in a variety of models.

weight_func: The type of kernel function that weights the distances between samples (e.g. in a K-near neighbors model).

surv_dist: the statistical distribution of the data in a survival analysis model (e.g. parsnip::surv_reg()) .

Laplace: the Laplace correction used to smooth low-frequency counts.

neighbors: a parameter for the number of neighbors used in a prototype model.