Parameter objects related to parametric models.

These are objects that can be used for modeling, especially in conjunction with the parsnip package.

dropout

epochs

activation

mixture

penalty

rbf_sigma

prod_degree

num_terms

num_comp

cost

scale_factor

margin

degree

deg_free

hidden_units

batch_size

Format

An object of class quant_param (inherits from param) of length 7.

Value

Each object is generated by either new_quant_param or new_qual_param.

Details

These objects are pre-made parameter sets that are useful when the model is based on some type of slope/intercept model.

penalty: The total amount of regularization used. This is used by parsnip::linear_reg() and parsnip::logistic_reg() with glmnet models.

mixture: the proportion of L1 regularization in the model. (parsnip::linear_reg() and parsnip::logistic_reg())

dropout: the parameter dropout rate. (parsnip:::mlp())

epochs: the number of iterations of training. (parsnip:::mlp())

activation: the type of activation function between network layers. (parsnip:::mlp())

hidden_units: the number of hidden units in a network layer. (parsnip:::mlp())

batch_size: the mini-batch size for neural networks.

rbf_sigma: the sigma parameters of a radial basis function.

cost: a cost value for SVM models.

scale_factor: the polynomial and hyperbolic tangent kernel scaling factor.

margin: the SVM margin parameter (e.g. epsilon in the insensitive-loss function for regression).

degree: the polynomial degree.

prod_degree: the number of terms to combine into interactions. A value of 1 implies an additive model. Useful for MARS models.

num_terms: a nonspecific parameter for the number of terms in a model. This can be used with models that include feature selection, such as MARS.

num_comp: the number of components in a model (e.g. PCA or PLS components).

deg_free: a parameter for the degrees of freedom.