Amount of regularization/penalization

A numeric parameter function representing the amount of penalties (e.g. L1, L2 etc) in regularized models.

penalty(range = c(-10, 0), trans = log10_trans())

Arguments

range
A two-element vector holding the defaults for the smallest and largest possible values, respectively. Note that these are in transformed units.

trans
A trans object from the scales package, such as scales::log10_trans() or scales::reciprocal_trans(). If not provided, the default is used which matches the units used in range. If no transformation, NULL.

Details

This parameter is used for regularized or penalized models such as parsnip::linear_reg(), parsnip::logistic_reg(), and others.

Examples

penalty()

#> Amount of Regularization  (quantitative)
#> Transformer:  log-10 
#> Range (transformed scale): [-10, 0]