Create grids of tuning parameters

Random and regular grids can be created for any number of parameter objects.

regular_grid(..., levels = 3, original = TRUE)

random_grid(..., size = 5, original = TRUE)

Arguments

...
One or more param objects (such as mtry() or regularization()). None of the objects can have unknown() values in the parameter ranges or values.

levels
An integer for how many value of each parameter will be used to make the regular grid? levels can be a single integer or a vector of integers that is the same length as the number of parameters in ....

original
A logical: should the parameters be in the original units or in the transformed space (if any)?

size
A single integer for the total number of parameter values returned for the random grid.

Value

A tibble with an additional class for the type of type of grid ("regular_grid" or "random_grid"). There are columns for each parameter and a row for every parameter or parameter combination.

Examples

# Will fail due to unknowns:
# regular_grid(mtry, min_n)

regular_grid(regularization, mixture)

#> # A tibble: 9 x 2
#>   regularization mixture
#>            <dbl>   <dbl>
#> 1   0.0000000001     0  
#> 2   0.00001          0  
#> 3   1                0  
#> 4   0.0000000001     0.5
#> 5   0.00001          0.5
#> 6   1                0.5
#> 7   0.0000000001     1  
#> 8   0.00001          1  
#> 9   1                1  

regular_grid(regularization, mixture, levels = c(3, 4))

#> # A tibble: 12 x 2
#>    regularization mixture
#>             <dbl>   <dbl>
#>  1   0.0000000001   0    
#>  2   0.00001        0    
#>  3   1              0    
#>  4   0.0000000001   0.333
#>  5   0.00001        0.333
#>  6   1              0.333
#>  7   0.0000000001   0.667
#>  8   0.00001        0.667
#>  9   1              0.667
#> 10   0.0000000001   1    
#> 11   0.00001        1    
#> 12   1              1    

random_grid(regularization, mixture)

#> # A tibble: 5 x 2
#>   regularization mixture
#>            <dbl>   <dbl>
#> 1 0.000000000642   0.466
#> 2 0.0220           0.498
#> 3 0.000102         0.290
#> 4 0.00000000373    0.733
#> 5 0.000000000119   0.773