Parameter functions related to text analysis.

These are parameter generating functions that can be used for modeling, especially in conjunction with the textrecipes package.

weight(range = c(-10, 0), trans = log10_trans())

weight_scheme(values = values_weight_scheme)

token(values = values_token)

max_times(range = c(1L, as.integer(10^5)), trans = NULL)

min_times(range = c(0L, 1000L), trans = NULL)

max_tokens(range = c(0L, as.integer(10^5)), trans = NULL)

Arguments

range
A two-element vector holding the smallest and largest possible values, respectively. If these cannot be set when the parameter is created, unknown() can be used instead of a particular value. If a transformation is specified, these should be in the transformed units. The default is an educated guess at the range of each parameter.

trans
A trans object from the scales package, such as scales::log10_trans() or scales::reciprocal_trans(). If not provided, the default is used which matches the units used in range. If no transformation, NULL.

values
A vector of possible parameter values. For qualitative parameters, character and logical defaults have been set, but can be modified as needed.

Value

Each object is generated by either new_quant_param() or new_qual_param().

Details

These functions generate parameters that are useful in a variety of text models.

min_times(), max_times(): The frequency of word occurances for removal. (See textrecipes::step_tokenfilter()).

max_tokens(): The number of tokens that will be retained. (See textrecipes::step_tokenfilter).

weight(): A parameter for "double normalization" when creating token counts. (See textrecipes::step_tf()).

weight_scheme(): The method for term frequency calculations. (See textrecipes::step_tf()).

token(): The type of token. (See textrecipes::step_tokenize()).