Create grids of tuning parameters

Random and regular grids can be created for any number of parameter objects.

grid_regular(..., levels = 3, original = TRUE)

grid_random(..., size = 5, original = TRUE)

Arguments

...
One or more param objects (such as mtry() or penalty()). None of the objects can have unknown() values in the parameter ranges or values.

levels
An integer for the number of values of each parameter to use to make the regular grid. levels can be a single integer or a vector of integers that is the same length as the number of parameters in ....

original
A logical: should the parameters be in the original units or in the transformed space (if any)?

size
A single integer for the total number of parameter value combinations returned for the random grid.

Value

A tibble with an additional class for the type of grid ("grid_regular" or "grid_random"). There are columns for each parameter and a row for every parameter combination.

Examples

# Will fail due to unknowns:
# grid_regular(mtry(), min_n())

grid_regular(penalty(), mixture())

#> # A tibble: 9 x 2
#>        penalty mixture
#>          <dbl>   <dbl>
#> 1 0.0000000001     0  
#> 2 0.00001          0  
#> 3 1                0  
#> 4 0.0000000001     0.5
#> 5 0.00001          0.5
#> 6 1                0.5
#> 7 0.0000000001     1  
#> 8 0.00001          1  
#> 9 1                1  

grid_regular(penalty(), mixture(), levels = c(3, 4))

#> # A tibble: 12 x 2
#>         penalty mixture
#>           <dbl>   <dbl>
#>  1 0.0000000001   0    
#>  2 0.00001        0    
#>  3 1              0    
#>  4 0.0000000001   0.333
#>  5 0.00001        0.333
#>  6 1              0.333
#>  7 0.0000000001   0.667
#>  8 0.00001        0.667
#>  9 1              0.667
#> 10 0.0000000001   1    
#> 11 0.00001        1    
#> 12 1              1    

grid_random(penalty(), mixture())

#> # A tibble: 5 x 2
#>         penalty mixture
#>           <dbl>   <dbl>
#> 1 0.00000000804  0.155 
#> 2 0.000000545    0.0999
#> 3 0.907          0.955 
#> 4 0.000165       0.717 
#> 5 0.0000000876   0.521